Update runtime to 25.08
This replaces p7zip with 7zip so that yasm is no longer needed. The
version of libjpeg used failed to build due to requiring an older cmake
version. It seems libjpeg is included in the org.freedesktop.Platform
runtime, so it should be unnecessary. It also doesn't seem to be a
dependency of winetricks anyway.

See org.winehq.Wine flatpak, as it includes winetricks:
https://github.com/flathub/org.winehq.Wine/blob/e751ada8fa620a8b5a9e533e7ceb5eb9cd057cd1/org.winehq.Wine.yml#L357-L372